Added a simple USB driver
The rgb_led modules create a rgb_led device in /dev.
To set the rgb values 3 bytes are expected to be written in plain text in
hexadecimal as following:

$ echo "a8f801" >> /dev/rgb_led1
In that case:
r = 0x01
g = 0xf8
b = 0xa8

Things are far from being bullet proof in that state but this is a first
working iteration.